Save progress

+++ BubbleHeroEngine/BubbleHeroEngine/Others/Settings.swift
@@ -17,5 +17,5 @@ class Settings {
     /// The allowed number of bubbles on even rows.
     static let cellPerRow = 12
     /// The number of types of bubbles in the game currently.
-    static let numOfTypes = 4
+    static let numOfTypes = UInt32(4)
 }
